Utils restructure.
