+++ tests/end_to_end/main.py
@@ -1,61 +1,8 @@
 import argparse
 import os
-from typing import Tuple
 
-import depthai as dai
 import pytest
-from utils import get_model_slugs_from_zoo
-
-
-def parse_model_slug(full_slug) -> Tuple[str, str]:
-    if ":" not in full_slug:
-        raise NameError(
-            "Please provide the model slug in the format of 'model_slug:model_version_slug'"
-        )
-    model_slug_parts = full_slug.split(":")
-    model_slug = model_slug_parts[0]
-    model_version_slug = model_slug_parts[1]
-
-    return model_slug, model_version_slug
-
-
-def find_slugs(parser: str):
-    relevant_slugs = []
-    print("Downloading NN archives from the ZOO")
-    slugs = get_model_slugs_from_zoo()
-    n = len(slugs)
-    for i, slug in enumerate(slugs):
-        if "dm-count" in slug:
-            print(f"[{i+1}/{n}] Skipping {slug} ...")
-            continue
-        print(f"[{i+1}/{n}] Downloading {slug} ...")
-        model_slug, version_slug = parse_model_slug(slug)
-        model_desc = dai.NNModelDescription(
-            modelSlug=model_slug, modelVersionSlug=version_slug, platform="RVC2"
-        )
-        nn_archive_path = None
-        try:
-            nn_archive_path = dai.getModelFromZoo(model_desc)
-        except Exception:
-            model_desc = dai.NNModelDescription(
-                modelSlug=model_slug, modelVersionSlug=version_slug, platform="RVC4"
-            )
-            try:
-                nn_archive_path = dai.getModelFromZoo(model_desc, useCached=True)
-            except Exception as e:
-                raise ValueError(f"Couldn't find model {slug} in the ZOO") from e
-
-        try:
-            nn_archive = dai.NNArchive(nn_archive_path)
-            for head in nn_archive.getConfig().model.heads:
-                if head.parser == parser:
-                    relevant_slugs.append(slug)
-                    break
-        except Exception as e:
-            print(e)
-
-        print(f"[{i+1}/{n}] Successfully downloaded {slug}!")
-    return relevant_slugs
+from utils import find_slugs, get_model_slugs_from_zoo
 
 
 def main():
